6.16
The algorithm given in the book actually doesn't work here -- when
implemented as written, men will end up standing next to each other when
moving one space to the right would leave a space between them. This is
a consequence of being unable to find a "true middle" of a run of even
length.

Ideally, a proper algorithm for maximizing restroom stall distance would
take in account that a man would start on the furthest extreme, the
second man would take the other extreme, and the third man would take
the central position. Subsequent calculations should follow from this -
find the maximum distance from occupants on your left *and* your right,
instead of the center of the longest unoccupied space.